bugfix/ authenticated user is acquired by loadByLoginProperty() method.
