+++ src/main/java/com/cinetime/controller/business/CinemaController.java
@@ -31,7 +31,7 @@ public class CinemaController {
     public ResponseMessage<Page<CinemaSummaryResponse>> listCinemas(
             @RequestParam(required = false) Long cityId,
             @RequestParam(required = false) Boolean specialHall,
-            @PageableDefault(page = 0, size = 10, sort = "title", direction = Sort.Direction.ASC)Pageable pageable) {
+            @PageableDefault(page = 0, size = 10, sort = "name", direction = Sort.Direction.ASC)Pageable pageable) {
 
      return cinemaService.listCinemas(cityId, specialHall, pageable);
